refactor: Split ui_components.py into smaller modules
